requirements.txt
annotated-types==0.7.0
anvil-app-server==1.12.0
anvil-uplink==0.5.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
openai==1.30.5
progressbar2==4.4.2
psutil==5.9.8
pychrome==0.2.4
pydantic==2.7.2
pydantic_core==2.18.3
python-dotenv==1.0.1
python-utils==3.8.2
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
websocket-client==1.8.0
ws4py==0.5.1
